Move ActiveWindowSet and implementations to runners-core
