+++ runners/core-java/src/main/java/org/apache/beam/runners/core/ActiveWindowSet.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.sdk.util;
+package org.apache.beam.runners.core;
 
 import com.google.common.annotations.VisibleForTesting;
 import java.util.Collection;
